pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "pycantoether"
description = "This is bridge between NMEA2000 and TCP server."
authors = [
    {name = "Alexey Matveev", email = "a.matveev@gmail.com"},
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
license = "Apache-2.0"
license-files = ["LICEN[CS]E*"]
dynamic = ["version", "readme"]

dependencies = [
    "python-can",
    "python-can[gs_usb]",
    "serial",
    "pyserial",
    "pyusb",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
]

[project.urls]
documentation = "https://github.com/shizacat/n2k-can2ether"
repository = "https://github.com/shizacat/n2k-can2ether"

[project.scripts]
pycantoether = "pycantoether.server:main"

[tool.setuptools.dynamic]
version = {attr = "pycantoether.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*",]
include = ["pycantoether", "pycantoether.*"]
namespaces = false
